requirements.txt
injector>=0.19.0
tiktoken>=0.3.3
pyyaml>=6.0
typing-extensions>=4.4.0
dataclasses>=0.8
uuid>=1.30

